feat: step5. 리팩토링
- View와 main method간 의존성 제거
- Controller 사용
